+++ CHANGELOG.md
@@ -2,6 +2,7 @@ development:
   * add SQL logger to development environment
   * generate public dir when running generator
   * use tux as Sinatra console
+  * when has no option, melodiest command will use --help option
 
 0.3.0: